+++ _build/gpm_resolvers/gpm.resolve.tables.php
@@ -22,11 +22,9 @@ if ($object->xpdo) {
             $modx->addPackage('glossary', $modelPath, null);
 
 
-
             $manager = $modx->getManager();
 
             $manager->createObjectContainer('Term');
-
 
             break;
     }