+++ core/components/glossary/docs/changelog.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2018-01-16
+### Added
+- Don't replace Glossary terms in definable html tags
+- Don't replace Glossary terms in all html tag attributes
+
 ## [2.3.0] - 2018-12-15
 ### Added
-- Don't replace crosslinks in definable html attributes
+- Don't replace Glossary terms in definable html tag attributes
 - System settings for sectionsStart and sectionsEnd
 - Edit Glossary system settings in the custom manager page